% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStepFunction.R
\name{addstepfun}
\alias{addstepfun}
\title{Add a step functions to an existing plot}
\usage{
addstepfun(x, y, lwd = par("lwd"), lty = par("lty"), col = "black",
  verticals = TRUE, do.points = FALSE, cex.points = par("cex"),
  pch = par("pch"), cadlag = TRUE)
}
\arguments{
\item{x, y}{vectors of x- and y- coordinates to specify the domain, jump points
and jump heights in the step function.}

\item{lwd, lty, col, cex.points, pch}{plot options for line width, line type,
color, size of points and type of points; see \code{\link{par}} for more details.}

\item{verticals}{a boolean variable that indicates whether vertical lines
should be added at all jump points.}

\item{do.points}{a boolean variable that indicates whether points
should be plotted at each jump point. The position of the points
depend on the choice for \code{cadlag}}

\item{cadlag}{a boolean variable to choose whether the step function should
be cadlag (right continous with left limits) or ladcag (left continous with
right limits). Has no effect if \code{do.points=FALSE}.}
}
\description{
A function to add a step function to the current plot. This function is
similar to the default \code{plot} of a function generated by
\code{\link{stepfun}} with the difference that the domain of the function
can be specified. More precisely, the first and last entries of the vector
\code{x} determine the range of the domain, i.e. the leftmost line in the
plot starts at \code{x[1]} and the rightmost line ends at \code{x[N]} where
\code{N} denotes the length of \code{x}. The x-coordinates of the jump points
are given by \code{x[2:N]} wheres the y-coordinate of each line is given
by the vector \code{y}. In particular \code{y} has to be a vector of length
\code{N-1}.
}
\seealso{
\code{\link{plotstep}}
}
