% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{alpha1_default}
\alias{alpha1_default}
\title{Default choice for \code{alpha1Func}}
\usage{
alpha1_default(alpha, K)
}
\arguments{
\item{alpha, K}{input arguments (overall prediction error and number of repetitions)}
}
\value{
the choice for alpha1 in predictions
}
\description{
The default choice for the function \code{alpha1Func} used in simulations
}
\seealso{
\code{\link{createSimulationFunction}, \link{createBatch}}
}
