% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoExp.R
\name{checkHypoexp}
\alias{checkHypoexp}
\title{Function to check for numeric instabilities}
\usage{
checkHypoexp(rates, precision = 0.001)
}
\arguments{
\item{rates}{a vector of rates used in the hypoexponential distribution}
}
\value{
a boolean value to indicate whether \code{sdprisk::phypoexp} seems 
numerically stable
}
\description{
A boolean function that checks whether the fast implementation in
\code{sdprisk::phypoexp} has severe numerical instabilities. To this end,
the function evaluates the sum of all coefficients in the cdf, which should 
add up to one. If the sum deviates from one by more than the prespecified 
\code{precision} value, the function returns \code{FALSE} and 
\code{sdprisk::phypoexp} should not be used. Note that even in the case
where the return value is \code{TRUE}, other kinds of numeric instabilites 
can occur. In situations with numeric instabilities the more stable 
implementation in \code{\link{phypoexp2}} should be used.
}
\seealso{
\code{\link{phypoexp2}}
}
