% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{coeff_basq}
\alias{coeff_basq}
\title{Coefficients in the cdf of the hypoexponential distribution}
\usage{
coeff_basq(theta, i, I0, Ic, I, s)
}
\arguments{
\item{theta, I, s}{parameter choices for the model}

\item{i}{index of the coefficient that is returned by the function}

\item{I0, Ic}{start and end in terms of broken wires for the rates}
}
\value{
the coefficient a_i(theta) of the cdf of the hypoexponential 
distribution
}
\description{
Coefficients \code{a_i(theta)} appearing in the cdf of the 
hypoexponential distribution.
}
