% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{coeff_vec_basq}
\alias{coeff_vec_basq}
\title{Coefficient vector in the cdf of the hypoexponential distribution}
\usage{
coeff_vec_basq(theta, I0, Ic, I, s)
}
\arguments{
\item{theta, I0, Ic, I, s}{parameter choices in \code{\link{coeff_basq}}}
}
\value{
vector of values given by \code{\link{coeff_basq}} for \code{i=I0:Ic}
}
\description{
Vector of coefficients computed via \code{\link{coeff_basq}}.
}
