% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConSets.R
\name{ConSet}
\alias{ConSet}
\title{Computation of confidence sets}
\usage{
ConSet(data, method = "3depth", alpha = 0.05, BasquinModel = TRUE,
  linkfun = NULL, gradlinkfun = NULL, grid_xlim = c(20, 35),
  grid_ylim = c(0, 5), mesh_width = 0.05, checkFullGrid = FALSE,
  autoExpandGrid = TRUE, expandX = 5, expandY = 5,
  maxExpansion_xlim = c(0, 100), maxExpansion_ylim = c(0, 50))
}
\arguments{
\item{data}{a list containing the observed data. The list has to be given in
format \code{list(w=waitingtimes, s=stresses, Iv=maxFailures, I=totalWires)}}

\item{method}{method to obtain a confidence set. Valid choices are "3depth", "wald" and "LR"}

\item{alpha}{the setting for the confidence level. The output has an (asymptotic) \code{(1-alpha)} confidence level.}

\item{BasquinModel}{a boolean variable indicating whether the default Basquin model
should be used. If set \code{TRUE} then \code{linkfun} and \code{gradlinkfun} will be changed
to the corresponding functions in the Basquin model.}

\item{linkfun}{the linkfunction used for the model}

\item{gradlinkfun}{the derivative of \code{linkfun} with respect to the parameter \eqn{\theta}}

\item{grid_xlim}{the horizontal grid limits for the first coordinate of the parameter \eqn{\theta}}

\item{grid_ylim}{the vertical grid limits for the second coordinate of the paramter \eqn{\theta}}

\item{mesh_width}{the mesh widh of the grid}

\item{checkFullGrid}{a boolean variable. If \code{TRUE} the algorithm will scan the entire grid
to compute the confidence set. If \code{FALSE} the algorithm will stop once
a full connected component of the confidence set is found.}

\item{autoExpandGrid}{a boolean variable. If \code{TRUE} the grid will be expanded
if the algorithm could not find a full connected component of the confidence set.
The grid will be expanded in all direction if no point in the set was found before.
Otherwise, expansion occur in the directions where the boundary of the current part
of the set overlaps with the margin of the grid.}

\item{expandX, expandY}{the horizontal and vertical increments if the grid is
expanded according to \code{autoExpandGrid}}

\item{maxExpansion_xlim, maxExpansion_ylim}{the maximal horizontal  and vertical
limit that the grid may expand to.}
}
\value{
a matrix with three rows and a variable number of columns. Each column
\eqn{(x,y,z)^\top} represents the leftmost point (x,z) and the rightmost point (y,z)
of the confidence set on the horizontal level \eqn{\theta_2=z}.
}
\description{
Lists the boundary points of confidence sets computed for two-dimensional
parameters. Possible methods for the computation are the robust 3-depth method
and the classical methods based on Wald's test and the Likelihood ratio test.
The computation is based on a grid search in a specified region. The region
is expanded automatically if necessary. This auto expansion can be turned off
by setting \code{autoExpandGrid} to \code{FALSE}.
}
\examples{
stresses <- c(200,100)
max_failures <- c(7,3)
totalWires <- 35
rates <- h_basq(theta=c(28,3), x=c(35*200/(35-0:6), 35*100/(35-0:2)))
waitingtimes <- rexp(10, rate=rates)
data <- list(w=waitingtimes, s=stresses, Iv=max_failures, I=totalWires)
ConSet(data, method="3depth")
ConSet(data, method="wald")

}
\seealso{
\code{\link{GridSearch}, \link{ConSet2d_wald}}
}
