% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{createBatch}
\alias{createBatch}
\title{Create a registry for the \code{BatchJobs} package}
\usage{
createBatch(id, seedlist = 1:200, setSeeds = TRUE, n_sims = 50,
  K_range = 1:20, theta = c(28, 3), block_s = c(200, 100, 80),
  block_Iv = c(6, 6, 6), block_I = 10, pred_s = 60, pred_start = 4,
  pred_pos = 6, pred_I = 10, alpha = 0.1,
  alpha1Func = alpha1_default, distr = rexp, withOutliers = FALSE,
  LowerOutlierChance = 0.1, LowerOutlierFunc = function(x) {    
  return(1) }, UpperOutlierChance = 0.1, UpperOutlierFunc = function(x)
  {     return(10 * x) }, methods = c("3depth", "delta", "wald",
  "naive"))
}
\arguments{
\item{id}{id of the registry; see \code{BatchJobs::makeRegistry}}

\item{seedlist}{a vector of seeds. The length of the vector also determines
the number of parallel jobs. The i-th jobs uses \code{seedlist[i]} as a seed}

\item{setSeeds}{boolean variable to choose whether seeds should be set or not}

\item{n_sims, K_range, theta, block_s, block_Iv, block_I, pred_s, pred_start, pred_pos, pred_I, alpha, alpha1Func, distr, withOutliers, LowerOutlierChance, LowerOutlierFunc, UpperOutlierChance, UpperOutlierFunc, methods}{parameters
used in \code{\link{createSimulationFunction}}. The i-th parallel job will
evaluate the simulation function at position \code{seedlist[i]}}
}
\value{
a registry based on \code{BatchJobs::makeRegistry}
containing the jobs for the parallel computation
}
\description{
This function creates a registry to perform parallel simulations on a computer
cluster (LiDoNg at TU Dortmund). The function requires the \code{BatchJobs}
package
}
\seealso{
\code{\link{createSimulationFunction}, BatchJobs::makeRegistry}
}
