% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{createSimulationFunction}
\alias{createSimulationFunction}
\title{Create a wrapper function for simulations}
\usage{
createSimulationFunction(n_sims, setSeeds = TRUE, K_range = 1:30,
  theta = c(28, 3), block_s = c(200, 100, 80), block_Iv = c(6, 6, 6),
  block_I = 35, pred_s = 60, pred_start = 4, pred_pos = 6,
  pred_I = 35, alpha = 0.1, alpha1Func = alpha1_default,
  distr = rexp, withOutliers = FALSE, LowerOutlierChance = 0.1,
  LowerOutlierFunc = function(x) {     return(1) },
  UpperOutlierChance = 0.1, UpperOutlierFunc = function(x) {    
  return(10 * x) }, methods = c("3depth", "delta", "wald", "naive"))
}
\arguments{
\item{n_sims}{the number of simulations (iterations of \code{simulateKRepetition})
the returned function should perform.}

\item{setSeeds}{a boolean variable to choose whether the returned function
should use its input value to set a seed.}

\item{K_range, theta, block_s, block_Iv, block_I, pred_s, pred_start, pred_pos, pred_I, alpha, alpha1Func, distr, withOutliers, LowerOutlierChance, LowerOutlierFunc, UpperOutlierChance, UpperOutlierFunc, methods}{parameters
used in each of the \code{n_sims} iterations of \code{simulateKRepetition}}
}
\value{
a function with a single integer as an argument. The function
starts a simulation based on the chosen parameters.
}
\description{
This function returns a function to start simulations. The simulation function
will have a single integer argument which either sets the seed
(if \code{setSeeds=TRUE}) or is a dummy argument with no effect.
}
\seealso{
\code{\link{simulateKRepetition}, \link{createBatch}}
}
