% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{DataSanityCheck}
\alias{DataSanityCheck}
\title{Function to check whether data is given in the correct format}
\usage{
DataSanityCheck(data = list(w = NULL, s = NULL, Iv = NULL, I = 35))
}
\arguments{
\item{data}{the data used for other function}
}
\value{
\code{invisible(NULL)}. Function produces an error if the data
has a wrong format
}
\description{
An auxilliary function that produces an error if \code{data} is not given 
in the correct form for subsequent functions.
}
