% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoExp.R
\name{dhypoexp2}
\alias{dhypoexp2}
\title{Density function of the hypoexponential distribution}
\usage{
dhypoexp2(x, rates)
}
\arguments{
\item{x}{the position at which the density is evaluated}

\item{rates}{a vector of rates used in the hypoexponential distribution}
}
\value{
the value of the density at \code{x}
}
\description{
An alternative calculation for the density function of the hypoexponential 
distribution based on a matrix representation. A faster but numerically 
more instable implementation for the cdf is given in \code{sdprisk::dhypoexp}
}
\seealso{
\code{\link{phypoexp2}, \link{qhypoexp2}}
}
