% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEstimation.R
\name{estML}
\alias{estML}
\title{Maximum likelihood estimation}
\usage{
estML(data, BasquinModel = TRUE, linkfun = NULL, optim_start = c(28,
  3), thetaLower = c(-Inf, Inf), thetaUpper = c(Inf, Inf),
  iterateOptim = TRUE, returnMLEonly = TRUE, na.rm = TRUE)
}
\arguments{
\item{data}{The experiment data given in a list with vectors \code{s,w,Iv,I}
representing the parameters in each experiment.}

\item{BasquinModel}{A boolean variable to indicate whether the Basquin model
is used. If \code{TRUE}, \code{linkfun} is automatically set to
\code{\link{h_basq}} and the more efficient function \code{\link{estML_basq}}
is used to compute the MLE.}

\item{linkfun}{the link function used in the model.}

\item{optim_start}{the parameter used as \code{par} in \code{\link{optim}}}

\item{thetaLower, thetaUpper}{restrictions to \code{theta} in the MLE
computation. The first coordinate of the estimator is restricted to be between
\code{thetaLower[1]} and \code{thetaUpper[1]} whereas the second coordinate
is contained in the interval (\code{thetaLower[1], thetaUpper[1]})}

\item{iterateOptim}{a boolean variable. If \code{TRUE} then \code{optim} is
used repeatedly until no improvement is made in the maximization process.}

\item{returnMLEonly}{a boolean variable. If \code{TRUE}, only the MLE
estimator for \code{theta} is returned. Otherwise, other infomation such
as the log-likelihood value at the MLE is returned as well.}
}
\value{
The MLE for \code{theta}
}
\description{
A function to compute the MLE for \code{theta} in model given by \code{linkfun}.
}
\seealso{
\code{\link{estML_basq}}
}
