% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{get_log_wij_preds}
\alias{get_log_wij_preds}
\title{Compute predictions in Figure 2}
\usage{
get_log_wij_preds(data = loadData(), alpha = 0.1, alpha1 = 1 - sqrt(1
  - alpha), alpha2 = alpha1, precision = 0.01)
}
\description{
Compute predictions in Figure 2
}
