% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{getbreakvec}
\alias{getbreakvec}
\title{Auxilliary function}
\usage{
getbreakvec(vec_a)
}
\arguments{
\item{vec_a}{a vector of integers}
}
\value{
another vector \code{b} with a total length of \code{sum(vec_a)}. 
The vector \code{b} is a concatenation of the vectors 
\code{0:(vec_a[1]-1)},...,\code{0:(vec_a[N]-1)} with \code{N=length(vec_a)}
}
\description{
An auxilliary function often used to compute a vector of rates in the 
Basquin model. The function expands a vector (a_1,...,a_N) of integers to
a longer vector (0,...,a_1-1,....,0,....,a_N-1).
}
\seealso{
\code{\link{getRates}}
}
