% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConBandsExp.R
\name{getConBands}
\alias{getConBands}
\title{Confidence band for exponential distribution}
\usage{
getConBands(alpha, n_sim = 1e+06, theta = c(28, 3), s, Iv, I,
  linkfun = h_basq)
}
\arguments{
\item{alpha}{the error in the confidence band.}

\item{n_sim}{the number of repetitions (simulations) on which the computation
of the confidence band is based. Increase for more precision, decrease to speed
up the computation. Default choice is one million.}

\item{theta, s, Iv, I, linkfun}{parameter choices in \code{\link{sampleStandQuants}}}
}
\value{
a list with three entries: \code{upper_band} and \code{lower_band}
are two vectores in which the ith entry contains the upper/lower bound for
the ith quantile of the exponential distribution. The third entry \code{alpha} is solely for
convenience and returns the \code{alpha} chosen as input.
}
\description{
Compute a simultaneous \code{(1-alpha)}-confidence band for the sample quantiles
in the Basquin model. The confidence band compares the rescaled waiting times
with the standard exponential distribution. This rescaling is done as in
\code{\link{sampleStandQuants}}. The confidence band is computed based on the
Kolmogorov-Smirnov test statistic, i.e. by computing a band which contains
all sample quantiles for which the sup distance between
their empirical distribution function and the standard exponential distribution
is below the corresponding \code{alpha}-quantile.
}
\seealso{
\code{\link{sampleStandQuants}}
}
