% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{getCovRate}
\alias{getCovRate}
\title{Coverage rates of prediction intervals}
\usage{
getCovRate(jumptimes, predInts)
}
\arguments{
\item{jumptimes}{a vector of length N storing the times that are to be 
predicted}

\item{predInts}{a N by 2 matrix in which \code{predInts[n,]} denotes the
prediction interval for \code{jumptimes[n]}}
}
\value{
a number between zero and one that represents the relative number
of points in \code{jumptimes} that are covered by the corresponding 
prediction interval
}
\description{
A computation of the coverage rate of prediction intervals given as a
matrix \code{predInts}. The values that need to be predicted are given as
a vector \code{jumptimes}
}
