% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getFigure1}
\alias{getFigure1}
\title{Reproduce Figure 1}
\usage{
getFigure1(data = loadData(), xlim = c(0, 750), ylim = c(0, 12),
  xlab = Fig1_xlab, ylab = Fig1_ylab, legend1_txt = Fig1_legend1_txt,
  legend2_txt = Fig1_legend2_txt, cex = 1, lwd = 1, cex.lab = cex,
  pt.cex = cex, y.intersp = 1)
}
\description{
Reproduce Figure 1
}
