% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getFigure2}
\alias{getFigure2}
\title{Reproduce Figure 2}
\usage{
getFigure2(n_sim = 1e+06, xlab = "Quantiles of Exp(1)",
  ylab = "Sample Quantiles", ylim = c(0, 15),
  main = "Q-Q-Plot with 90\%-confidence band")
}
\description{
Reproduce Figure 2
}
