% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getFigure3a}
\alias{getFigure3a}
\title{Reproduce Figure 3a}
\usage{
getFigure3a(xlim = c(0, 3), ylim = c(0, 150),
  legend_txt = Fig3a_legend_txt, main = "",
  ylab = "Load cycles in millions",
  xlab = "Number of broken tension wires", lwd = 1)
}
\description{
Reproduce Figure 3a
}
