% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getFigure3b}
\alias{getFigure3b}
\title{Reproduce Figure 3b}
\usage{
getFigure3b(xlim = c(0, 17), ylim = c(0, 30),
  legend_txt = Fig3b_legend_txt, main = "",
  ylab = "Load cycles in millions",
  xlab = "Number of broken tension wires", lwd = 1)
}
\description{
Reproduce Figure 3b
}
