% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getFigureA1}
\alias{getFigureA1}
\title{Reproduce Figure A1}
\usage{
getFigureA1(xlab = Fig3_xlab, ylab = Fig3_ylab,
  legend_txt = Fig3_legend_txt, ylim = c(7, 8.1), lwd = 2,
  dist_ints = 0.2, data = loadData(), preds = getPred19SB04(data))
}
\description{
Reproduce Figure A1
}
