% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{getIntLength}
\alias{getIntLength}
\title{Interval lengths of prediction intervals}
\usage{
getIntLength(predInts)
}
\arguments{
\item{predInts}{a N by 2 matrix in which \code{predInts[n,]} denotes the
prediction interval for \code{jumptimes[n]}}
}
\value{
the average length (mean) of the intervals in \code{predInts}
}
\description{
A computation of the average interval length of prediction intervals given 
as a matrix \code{predInts}.
}
