% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getPred19SB04}
\alias{getPred19SB04}
\title{Additional Predictions for the last failure of SB04}
\usage{
getPred19SB04(data = loadData(), alpha = 0.1, alpha1 = 1 - sqrt(1 -
  alpha), alpha2 = alpha1)
}
\description{
Additional Predictions for the last failure of SB04
}
