% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{getRates}
\alias{getRates}
\title{Function to compute rates}
\usage{
getRates(theta, s, Iv, I, linkfun = h_basq)
}
\arguments{
\item{theta, s, Iv, I}{model parameters used in \code{linkfun}}

\item{linkfun}{link function used for the rate. Default choice is 
\code{\link{h_basq}}}
}
\value{
vector of rates for the hypoexponential distribution.
}
\description{
A function to compute the rates of a hypoexponential distribution for a given
link function \code{linkfun}.
}
\seealso{
\code{\link{h_basq},\link{getbreakvec}}
}
