% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScore.R
\name{getScore}
\alias{getScore}
\title{Interval score of prediction intervals}
\usage{
getScore(jumptimes, predInts, alpha)
}
\arguments{
\item{jumptimes}{a vector of length N storing the times that are to be 
predicted}

\item{predInts}{a N by 2 matrix in which \code{predInts[n,]} denotes the
prediction interval for \code{jumptimes[n]}}

\item{alpha}{the alpha-error chosen for the prediction intervals}
}
\value{
the average interval score of the prediction intervals
}
\description{
A computation of the interval score by Gneiting and Raftery (2007) of 
prediction intervals given as a
matrix \code{predInts}. The values that need to be predicted are given as
a vector \code{jumptimes}.
}
