% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructFigures.R
\name{getTable1}
\alias{getTable1}
\title{Reproduce Table 1}
\usage{
getTable1()
}
\description{
Note that the leave-one-out cross validation is time consuming. Hence the
computation of the table can take around 30 minutes to finish.
}
