% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{grad2_coeff_vec_basq}
\alias{grad2_coeff_vec_basq}
\title{Derivative of the coefficient vector with respect to \code{theta[2]}}
\usage{
grad2_coeff_vec_basq(theta, I0, Ic, I, s)
}
\arguments{
\item{theta, i, I0, Ic, I, s}{parameter choice in \code{\link{coeff_vec_basq}}}
}
\value{
Derivative of \code{\link{coeff_vec_basq}} with respect to \code{theta[2]}
}
\description{
Derivative of \code{\link{coeff_vec_basq}} with respect to the second 
coordinate of \code{theta} evaluated at \code{theta}.
}
