% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{grad_coeff_basq}
\alias{grad_coeff_basq}
\title{Gradient of the coefficients in the cdf of the hypoexponential distribution}
\usage{
grad_coeff_basq(theta, i, I0, Ic, I, s)
}
\arguments{
\item{theta}{value of theta where the gradient is evaluated}

\item{i, I0, Ic, I, s}{parameter choice for the coefficient in \code{\link{coeff_basq}}}
}
\value{
value of the gradient at position \code{theta}
}
\description{
Gradient with respect to \code{theta} of the coefficient given in 
\code{\link{coeff_basq}}
}
