% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{grad_lambda_basq}
\alias{grad_lambda_basq}
\title{Gradient of the rate function (Basquin Model)}
\usage{
grad_lambda_basq(theta, I, s, i)
}
\arguments{
\item{theta, I, s}{parameters used in \code{\link{getRates}}. The missing 
variable \code{Iv} can be any value which is at least \code{i}.}

\item{i}{coordinate(s) of \code{\link{getRates}} which is/are considered.}
}
\value{
a vector (or a matrix if \code{length(i)>1}) stating the values 
of the gradient at position \code{theta}
}
\description{
The gradient of the \code{i}-th coordinate of the function 
\code{\link{getRates}} with respect to 
\code{theta} when using link function \code{\link{h_basq}}.
}
\seealso{
\code{\link{getRates},\link{h_basq}}
}
