% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{grad_qhypoexp_basq}
\alias{grad_qhypoexp_basq}
\title{Gradient of the quantile function of the hypoexponential distribution}
\usage{
grad_qhypoexp_basq(theta, alpha, I0, Ic, I, s)
}
\arguments{
\item{theta}{value of theta where the gradient is evaluated}

\item{alpha}{p-value at which the quantile function is evaluated}

\item{I0, Ic}{start and end in terms of broken wires for the rates}

\item{I, s}{parameter choice (max. number of wires and initial stress range)
in the unterlying experiment.}
}
\value{
Gradient of the quantile function with respect to \code{theta}
}
\description{
Gradient with respect to \code{theta} of the quantile function of the 
hypoexponential distribution with link function \code{\link{h_basq}}.
The calculation of the gradient is based on the implicit function theorem.
}
