% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConSets.R
\name{GridSearch}
\alias{GridSearch}
\title{Grid search for confidence sets}
\usage{
GridSearch(InSetFun, grid_xlim, grid_ylim, mesh_width, checkFullGrid)
}
\arguments{
\item{InSetFun}{a boolean function to represent the set}

\item{grid_xlim, grid_ylim, mesh_width, checkFullGrid}{options for the grid search.
Details can be found in \code{\link{ConSet}}}
}
\value{
a matrix with three rows and a variable number of columns.
Details can be found in \code{\link{ConSet}}
}
\description{
A grid search that finds points of a set on a given grid. The set is represented
by a boolean function \code{InSetFun} which returns true if and only if the input
is part of the set.
}
\seealso{
\code{\link{ConSet}, \link{GridSearchX}}
}
