% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConSets.R
\name{GridSearchX}
\alias{GridSearchX}
\title{Auxiliary function for grid search}
\usage{
GridSearchX(InSetFun, y, x_candidates)
}
\arguments{
\item{InSetFun}{a boolean function. Details can be found in \code{\link{GridSearch}}.}

\item{y}{a real number for the fixed y-coordinate considered in the set.}

\item{x_candidates}{a vector of x-coordinates which are tested during
grid search on level \code{y}.}
}
\value{
a vector of length two containing the minimal- and maximal x values
of points inside the set that have y-coordinate \code{y}. The function returns
\code{NULL} if none of the pointes listed in \code{x_candidates} belong to the set.
}
\description{
A function to find the minimal- and maximal x-coordinate of points in a set
for a fixed y coordinate. This function is used repeatedly to perform
\code{\link{GridSearch}}.
}
\seealso{
\code{\link{GridSearch}}
}
