% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{h_basq}
\alias{h_basq}
\title{Link function based on the Basquin Model}
\usage{
h_basq(theta, x)
}
\arguments{
\item{theta}{a vector of length two used as the parameter in the link function}

\item{x}{a vector of real numbers used as arguments in the link function.}
}
\value{
a vector with the same length as x.
}
\description{
Default choice for the link function in all predictions.
}
\seealso{
\code{\link{getRates}}
}
