% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{inv_cramer}
\alias{inv_cramer}
\title{Inverse matrix with Cramer's rule}
\usage{
inv_cramer(A)
}
\arguments{
\item{A}{a 2x2 matrix}
}
\value{
Inverse of \code{A}
}
\description{
Function to compute the inverse of a 2x2 matrix using Cramer's rule
}
