% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealDataAnalysis.R
\name{leaveOneOutData}
\alias{leaveOneOutData}
\title{Split data into observed and prediction for a leave-one-out analysis}
\usage{
leaveOneOutData(data, num_experiment, obs_in_experiment)
}
\arguments{
\item{data}{the entire data set as a list supported by
\code{\link{DataSanityCheck}}}

\item{num_experiment}{the number of the experiment that is to be predicted
in the leave-one-out analysis}

\item{obs_in_experiment}{the number of jumps that are observed
in experiment number \code{num_experiment} which do not need to be predicted}
}
\value{
a list which contains another list \code{observations} of observed
data in the leave-one-out analysis and the information \code{pred_initial_value}
and \code{pred_jumps} of the data that needs to be predicted.
}
\description{
A function to that split \code{data} into a list \code{observations} of
observed data and suitable information (\code{pred_initial_value,pred_jumps})
on data that is to be predicted. The observed data contains
all experiment except for (parts of) experiment number \code{num_experiment}.
The jump times up to index \code{obs_in_experiment} of experiment number
\code{num_experiment} are also added to the observed data in
\code{observations}. The remaining jump times are considered to be
part of the data that is to be predicted. In particular,
\code{pred_initial_value} equals either 0 (if \code{obs_in_experiment=0})
or is given as jump number \code{obs_in_experiment} of experiment number
\code{num_experiment}.
}
\seealso{
\code{\link{predNext_LOO}}
}
