% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealDataAnalysis.R
\name{loadData}
\alias{loadData}
\title{Load experiment data}
\usage{
loadData(mergeToList = TRUE, excludeSB06a = TRUE)
}
\arguments{
\item{mergeToList}{a boolean variable. If \code{TRUE}, the function returns
the data in form of a list that suits the format required in
\code{\link{DataSanityCheck}}}

\item{excludeSB06a}{a boolean variable to decide whether the results
from experiment SB06a should be removed from the list created when
\code{mergeToList=TRUE}}
}
\value{
The experiment data in form a list
}
\description{
A function to load the experiment results performed in the civil engineering
department of the TU Dortmund university. The data contains
the stress ranges \code{stress}, the times \code{jumptimes} when wires break
and the times \code{t_jm} between consecutive wire breaks in each experiment.
Each experiment is based on a concrete beam with I=35 wires.
The data in experiment SB06a was obtained after increasing the stress
range in SB06 since only one wire break was observed after a long running
time. We recommend to remove the data from SB06a since it is not comparable
to the other data.
}
