% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{logNormSampler}
\alias{logNormSampler}
\title{Generate a lognormal sample distribution}
\usage{
logNormSampler(adjustSDtoExp = TRUE, SDExpRatio = 1, sdFixed = 10)
}
\arguments{
\item{adjustSDtoExp}{boolean variable to indicate whether the standard
deviation of the resulting distribution should be given by \code{SDExpRatio}-times
the standard deviation of the exponential distribution or should take the
fixed value \code{sdFixed}}

\item{SDExpRatio}{the ratio between the standard deviations if \code{adjustSDtoExp=TRUE}.
Default choice is 1.}

\item{sdFixed}{the value for the standard deviation of the lognormal distribution
if \code{adjustSDtoExp=FALSE}}
}
\value{
a function with arguments \code{N} and \code{rates} that generates
\code{N} random variables with lognormal distribution.
}
\description{
Generates a sample distribution with two arguments \code{N} and \code{rates}.
The distribution generates \code{N} random variables with a lognormal
distribution. The parameters \code{mean} and \code{sd} of the underlying
normal distribution are chosen in the following way: \code{mean} is
chosen such that the median of the lognormal distribution coincides with the
median of the exponential distribution with \code{rate=rates}.
and normally distributed random variables with \code{mean} zero and a
standard deviation given by the parameter \code{sd}. If \code{adjustSDtoExp=TRUE}
then \code{sd} is chosen in such a way that ratio betwenn the standard
deviation of the lognormal distribution and the one of the exponential
distribution is equal to the chosen value \code{SDExpRatio}. Otherwise,
the standard deviation is set to be the fixed value \code{sdFixed}
}
\seealso{
\code{\link{sample_w}}
}
