% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{normed3depth}
\alias{normed3depth}
\title{3-depth test statistics}
\usage{
normed3depth(residuals)
}
\arguments{
\item{residuals}{a vector of residuals. Note that the vector can be replaced
by \code{sign(residuals)} since the 3-depth is only based on sign changes.}
}
\value{
the normalized 3-depth of \code{residuals}
}
\description{
A function to compute the normalized 3-depth of the residual vector
\code{residuals}. The normalized 3-depth is given by
\code{N(d3(residuals)-0.25)}, where \code{N} denotes the length of
\code{residuals}. The equivalent representation used in this function is
based on Equation (3.2) in Kustosz, Leucht, Mueller (2014).
}
