% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConBandsExp.R
\name{plotConBands}
\alias{plotConBands}
\title{Q-Q-plot with confidence bands}
\usage{
plotConBands(bands, alpha = bands$alpha, theo_q, sample_q,
  q_type = "l", col_band = "gray", fill_bands = TRUE,
  fill_col = "gray", add_diagonal = TRUE, lty_diag = 2,
  ylim = NULL, xlab = "Quantiles of Exp(1)",
  ylab = "Sample Quantiles", main = paste("Q-Q-Plot with ", 100 * (1 -
  alpha), "\%-confidence band", sep = ""))
}
\arguments{
\item{bands}{a list providing the confidence band, see \code{\link{getConBands}}.}

\item{alpha}{the error chosen for \code{bands}. Only used for the default
\code{main} of the plot}

\item{theo_q, sample_q}{two vectors storing the theoretical- and sample quantiles
for the QQ-plot.}

\item{q_type}{the type for the QQ-plot. Default is a linear interpolation of
the quantiles, see \code{type} in \code{\link{plot}} for more options.}

\item{col_band, fill_bands, fill_col}{options for plotting the confidence band.
\code{col_band} and \code{fill_col} provide the colors for the border of the
confidence band and the fill color of the bands. The fill color is only used if
\code{fill_bands} is \code{TRUE}.}

\item{add_diagonal, lty_diag}{If \code{add_diagonal=TRUE}, the diagonal \code{y=x}
is added to the plot. The line type of the diagonal can be adjusted via \code{lty_diag}.}

\item{ylim, xlab, ylab, main}{parameter choices in \code{\link{plot}}.
The default setting \code{ylim=NULL} yields an automated choice for \code{ylim}.}
}
\value{
\code{invisible(NULL)}.
}
\description{
Creates a Q-Q-plot of the theoretical quantiles \code{theo_q} and the sample
quantiles \code{sample_q}. Also adds a confidence band given as a list \code{bands}.
This confidence band can be computed via \code{\link{getConBands}}.
}
\seealso{
\code{\link{sampleStandQuants},\link{plot}}
}
