% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConSets.R
\name{plotConSet}
\alias{plotConSet}
\title{A plot function for confidence sets}
\usage{
plotConSet(set, add = FALSE, col = "blue", alpha = 0.05,
  xlab = expression(theta[1]), ylab = expression(theta[2]),
  xlim = c(0.99 * min(set[1:2, ]), 1.01 * max(set[1:2, ])),
  ylim = c(0.99 * min(set[3, ]), 1.01 * max(set[3, ])),
  main = paste(100 * (1 - alpha), "\% Confidence set for \\\\Theta", sep
  = ""))
}
\arguments{
\item{set}{a matrix with three rows and a variable number of columns representing
the confidence set. See the return value of \code{\link{ConSet}} for more details.}

\item{add, col, xlim, ylim, main, xlab, ylab}{plot options. See \code{\link{plot}} and \code{\link{par}}.}

\item{alpha}{the alpha value chosen in \code{\link{ConSet}}. The chosen value
will be shown in the default plot title.}
}
\value{
\code{NULL}
}
\description{
A simple plot function for confidence sets given in the return format of \code{\link{ConSet}}.
}
\seealso{
\code{\link{ConSet}}
}
