% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{plotSimResult}
\alias{plotSimResult}
\title{Plot simulation results}
\usage{
plotSimResult(simResult, type = c(1, 3),
  methods = dimnames(simResult)[[1]],
  ylabs = dimnames(simResult)[[2]][type],
  xlabs = rep("Repetitions of stress ranges 200, 100, 80", length(type)),
  ylims = NULL, alpha = 0.1, mSymbols = 1:length(methods), lwd = 2,
  cex = 0.5, inMillions = c(FALSE, TRUE))
}
\arguments{
\item{simResult}{a three-dimensional array produced by
the simulation function returned by \code{\link{createSimulationFunction}}
or via the parallel computation induced by \code{\link{createBatch}}}

\item{type}{a vector with entries in {1,2,3}. Type 1 refers
to a plot of the coverage rates, type 1 yields a plot of the interval
lengths and type 3 provides a plot of the interval scores.}

\item{methods}{a vector of names of methods used to produce the simulation
results. These names will appear in the legend of the graphic.}

\item{ylabs, xlabs}{a vector of labels for y- and x-axis for
each of the plots produced by the function (the total number of plots
is equal to the length of \code{type})}

\item{ylims}{an optional parameter to specify the \code{ylim} in each plot.
If \code{length(type)>1}, then \code{ylims} needs to be a matrix with two
columns in which the i-th row represents the \code{ylim} choice for
\code{type[i]}}

\item{mSymbols}{a vector of \code{pch} choices for the points in each of the
\code{methods}. Has to have the same length as \code{methods}}

\item{lwd, cex}{additional plot options for the line width and the point size.}

\item{inMillions}{a vector of booleans with the same length as \code{type}.
The i-th entry specifies whether the y-axis in plot for \code{type[i]} should
be in millions. If \code{TRUE}, each y-value is divided by \code{1e6}

used in \code{\link{createSimulationFunction}}. The i-th parallel job will
evaluate the simulation function at position \code{seedlist[i]}}
}
\description{
A function to plot the simulation  results
}
\seealso{
\code{\link{createSimulationFunction}, \link{createBatch}}
}
