% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStepFunction.R
\name{plotstep}
\alias{plotstep}
\title{Plot of a step functions}
\usage{
plotstep(x, y, xlim = c(x[1], x[length(x)]), ylim = c(y[1],
  y[length(y)]), xlab = "x", ylab = "y", main = "",
  lwd = par("lwd"), lty = par("lty"), col = "black",
  verticals = TRUE, do.points = FALSE, cex = par("cex"),
  cex.lab = cex, cex.points = cex, pch = par("pch"), cadlag = TRUE)
}
\arguments{
\item{x, y, lwd, lty, col, cex.points, pch, verticals, do.points, cadlag}{paramter
choices in \code{\link{addstepfun}}}

\item{xlim, ylim, xlab, ylab, main, cex, cex.lab}{further plot options; see \code{\link{plot}}
for more details.}
}
\description{
A wrapper function to create a new plot and perfom the \code{\link{addstepfun}}
function.
}
\seealso{
\code{\link{addstepfun}}
}
