% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{pred_delta}
\alias{pred_delta}
\title{Prediction intervals based on the delta-method}
\usage{
pred_delta(pred_I, pred_s, pred_start, pred_pos, pred_initial_value,
  alpha1, alpha2, obs_data, BasquinModel = TRUE, linkfun = NULL,
  gradlinkfun = NULL, auto_matrix_rep = FALSE)
}
\arguments{
\item{pred_I, pred_s}{model parameters s and I for the experiment that is to
be predicted}

\item{pred_start}{the first waiting time that is not observed in the
predicted experiment.}

\item{pred_pos}{a vector of indices indicating which waitingtimes are to be
predicted}

\item{pred_initial_value}{a number representing the last jumptime observed
in the new experiments before the prediction starts}

\item{alpha1}{error in the correction to the MLE of the asymptotic quantiles
used as prediction intervals}

\item{alpha2}{choice for the quantiles estimated via MLE. Up to corrections
made according to the choice of \code{alpha1}, the prediction interval
coincides with a naive \code{(1-alpha2)}-prediction interval obtained via
\code{\link{predHypoexp}} with \code{alpha=alpha2}.}

\item{BasquinModel}{a boolean variable to indicate whether the Basquin model
is used. Note that the current version of this package only support the
Basquin model.}

\item{linkfun, gradlinkfun}{linkfunction and its gradient if a model
other than the Basquin model is used. Note that this has no effect in
the current version since only the Basquin model is supported.}

\item{auto_matrix_rep}{a boolean variable used in \code{\link{predHypoexp}}}

\item{obsData}{observed data which the prediction is based on. The data
has to be formated as in \code{\link{DataSanityCheck}}}
}
\value{
a matrix of \code{(1-alpha1)(1-alpha2)}-prediction intervals
in which the i-th row stores the prediction for the jump time
at \code{pred_pos[i]}
}
\description{
A function to compute \code{(1-alpha1)(1-alpha2)}-prediction intervals based
on the delta-method.
}
\seealso{
\code{\link{predfail}}
}
