% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{predfail}
\alias{predfail}
\title{Function to produce prediction intervals based on several methods}
\usage{
predfail(pred_I, pred_s, pred_start = 1, pred_pos,
  pred_initial_value = 0, alpha1 = 0.05, alpha2 = 0.05,
  alpha = alpha1 + alpha2, auto_matrix_rep_hypoexp = FALSE,
  obs_data = list(w = NULL, s = NULL, Iv = NULL, I = NULL),
  BasquinModel = TRUE, linkfun = NULL, gradlinkfun = NULL,
  method = "3depth", grid_xlim = c(20, 35), grid_ylim = c(0, 5),
  mesh_width = 0.05, checkFullGrid = FALSE, autoExpandGrid = TRUE,
  expandX = 5, expandY = 5, maxExpansion_xlim = c(0, 100),
  maxExpansion_ylim = c(0, 50))
}
\arguments{
\item{pred_I, pred_s}{model parameters s and I for the experiment that is to
be predicted}

\item{pred_start}{the first waiting time that is not observed in the
predicted experiment.}

\item{pred_pos}{a vector of indices indicating which waitingtimes are to be
predicted}

\item{pred_initial_value}{a number representing the last jumptime observed
in the new experiments before the prediction starts}

\item{alpha1}{error in the confidence set (if "wald", "LR", or "3depth" are
chose as a method) or the correction to the MLE of the asymptotic quantiles
used as prediction intervals (if "delta" is chosen as a method);
see \code{\link{ConSet},\link{pred_delta}} for more details.}

\item{alpha2}{choice for \code{alpha} in \code{\link{predHypoexp}}
if any method other than "naive" is chosen; see
\code{\link{predUnionConSet}, \link{pred_delta}} for more details.}

\item{auto_matrix_rep_hypoexp}{choice for \code{aut_matrix_rep} in
\code{\link{predHypoexp}}}

\item{BasquinModel}{a boolean variable to indicate whether the Basquin model
is used.}

\item{linkfun, gradlinkfun}{linkfunction and its gradient if a model
other than the Basquin model is used.}

\item{method}{a method choice. Supported choices are "naive", "wald", "LR",
"3depth" and "delta".}

\item{grid_xlim, grid_ylim, mesh_width, checkFullGrid, autoExpandGrid, expandX, expandY, maxExpansion_xlim, maxExpansion_ylim}{parameter
choices in \code{\link{ConSet}}}

\item{obsData}{observed data which the prediction is based on.
The data should be a list that stores a vector of waiting times
in \code{obsData$w} and the experiment settings of the observed
data in \code{obsData$s}, \code{obsData$Iv}, \code{obsData$I};
see also \code{\link{DataSanityCheck}} for the data format check}
}
\value{
a matrix of prediction intervals in which the i-th row stores the
prediction for the jump time at \code{pred_pos[i]}
}
\description{
A function to compute either \code{(1-alpha)}-prediction intervals
based on the naive approach or \code{(1-alpha1)(1-alpha2)}-prediction
intervals based on either one of the confidence set based approaches
(Wald, LR, 3-depth) or the delta-method.
}
\seealso{
\code{\link{predHypoexp}, \link{pred_delta}, \link{predUnionConSet}, \link{ConSet}}
}
