% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoExp.R
\name{predHypoexp}
\alias{predHypoexp}
\title{Prediction intervals for the hypoexponential distribution}
\usage{
predHypoexp(alpha, rates, use_matrix_rep = TRUE,
  auto_matrix_rep = TRUE)
}
\arguments{
\item{alpha}{the alpha-error in the prediction}

\item{rates}{a vector of rates used in the hypoexponential distribution}

\item{use_matrix_rep}{a boolean variable to choose whether 
\code{\link{qhypoexp2}} or \code{sdprisk::qhypoexp} should be used to compute
quantiles. The variable is overwritten by an automatic choice if 
\code{auto_matrix_rep=TRUE}}

\item{auto_matrix_rep}{a boolean variable. If \code{TRUE}, the choice 
for \code{use_matrix_rep} is done automatically by checking for numeric 
instabilities via \code{\link{checkHypoexp}}}
}
\value{
a \code{(1-alpha)}-prediction interval
}
\description{
A function that returns a \code{(1-alpha)}-prediction interval for a 
hypoexponential distribution with given \code{rates}. The prediction 
interval is given as a vector containing the \code{alpha/2}-quantile
and the \code{(1-alpha/2)}-quantile of the distribution.
}
\seealso{
\code{\link{qhypoexp2}}
}
