% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{predImat}
\alias{predImat}
\title{Estimation for the information matrix}
\usage{
predImat(thetahat, s, Iv, I, linkfun, gradlinkfun, BasquinModel)
}
\arguments{
\item{thetahat}{The MLE for theta. Can be computed via \code{\link{estML}}}

\item{s, Iv, I}{parameter settings in the experiment}

\item{linkfun, gradlinkfun}{Linkfunction and its gradient (with respect to
theta) for the model. Only needs to be specified if \code{BasquinModel=FALSE}}

\item{BasquinModel}{a boolean variable. If \code{TRUE} the Basquin model is
used. Otherwise, the model is specified via the choice of \code{linkfun}}
}
\value{
The estimated information matrix (multiplied by the factor
\code{N=sum(Iv)})
}
\description{
A function to compute the estimated information matrix for a dataset. Note
that the resulting matrix is in a format suited for further calculations.
In order to transform the return value to a properly scaled estimator for
the information matrix, a rescaling by the factor \code{1/N} is required,
where \code{N} denotes the number of observations, i.e.\code{N=sum(Iv)}.
}
\seealso{
\code{\link{predImat_basq}}
}
