% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasquinModel.R
\name{predImat_basq}
\alias{predImat_basq}
\title{Prediction of information matrix (Basquin Model)}
\usage{
predImat_basq(s, Iv, I)
}
\arguments{
\item{s}{vector of stress ranges in the experiments}

\item{Iv}{vector of total number of broken wires observed in each experiment}

\item{I}{total number of wires in each experiment}
}
\value{
a matrix estimating the Fisher information
}
\description{
The simplified formula to predict the information matrix when the Basquin 
model is used.
}
\seealso{
\code{\link{predImat}}
}
