% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealDataAnalysis.R
\name{predNext_LOO}
\alias{predNext_LOO}
\title{Perform a leave-one-out analysis on data}
\usage{
predNext_LOO(data, future_steps = 1, alpha = 0.1, alpha1 = 1 - sqrt(1
  - alpha), alpha2 = 1 - sqrt(1 - alpha), methods = c("3depth",
  "delta", "wald", "LR", "naive"))
}
\arguments{
\item{data}{the entire data set as a list supported by
\code{\link{DataSanityCheck}}}

\item{future_steps}{an integer to determine the index difference between
observations and the predicted jump time, i.e. the predicted random variable
will be a sum of \code{future_steps} exponentially distributed random variables}

\item{alpha, alpha1, alpha2}{error choices in the predictions; see
\code{\link{predfail}} for more details.}

\item{methods}{a vector of methods supported by \code{\link{predfail}}.}
}
\value{
a list which contains a matrix \code{result} and an integer
\code{n_observations}. The matrix result contains the average score, coverage
rate and interval length in each of the methods, averaged over all predictions
made during the LOO. the integer \code{n_observations} counts the total number
of predictions made during the LOO.
}
\description{
A function that performs a leave-one-out analysis (LOO) on \code{data}.
The for \code{future_steps} determines whether predictions
should be made for waiting times/immediate successors of jump times
(if \code{future_steps=1}) or for a sum of the next \code{future_steps}
waiting times after each jump time.
}
\seealso{
\code{\link{predfail},\link{leaveOneOutData}}
}
