% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{predUnionConSet}
\alias{predUnionConSet}
\title{Prediction intervals based on confidence sets}
\usage{
predUnionConSet(pred_I, pred_s, pred_start = 1, pred_pos,
  pred_initial_value = 0, alpha = 0.05, ConSet, linkfun = h_basq,
  approx = dim(ConSet)[2], auto_matrix_rep = FALSE)
}
\arguments{
\item{pred_I, pred_s}{model parameters s and I for the experiment that is to
be predicted}

\item{pred_start}{the first waiting time that is not observed in the
predicted experiment.}

\item{pred_pos}{a vector of indices indicating which waitingtimes are to be
predicted}

\item{pred_initial_value}{a number representing the last jumptime observed
in the new experiments before the prediction starts}

\item{alpha}{the error for each prediction interval}

\item{ConSet}{the confidence set in the data format provided by the function
\code{\link{ConSet}}}

\item{linkfun}{the linkfunction used for the model}

\item{approx}{an additional parameter to reduce computation time by
adding another approximation error. instead of considering all columns in
\code{ConSet}, only a number of \code{approx} equidistant ones is used.}

\item{auto_matrix_rep}{a boolean variable used in \code{\link{predHypoexp}}}
}
\value{
the union(s) over all (1-\code{alpha})-prediction intervals taken over
all parameters in \code{ConSet}. The result is given as a matrix with
2 columns in which the i-th row stores the prediction for the jump time
at \code{pred_pos[i]}
}
\description{
A function that transforms a confidence set \code{ConSet} with error
\code{alpha2} to (a matrix of) (1-\code{alpha})(1-\code{alpha2})-prediction
intervals for waiting times with indices \code{pred_pos} in an
experiment with parameters \code{pred_I} and \code{pred_s}.
If \code{pred_star>1}, the first \code{pred_start}-1 waiting times of
the new experiments are assumed to be part of the observed input. In this
case, the last jumptimes (i.e. the sum of waiting times 1,...,
\code{pred_start}-1) needs to be provided as another variable
\code{pred_initial_value}.
}
\seealso{
\code{\link{ConSet}, \link{predfail}}
}
