% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntervals.R
\name{q3depth}
\alias{q3depth}
\title{Quantiles for 3-depth}
\usage{
q3depth(p, N = 101, useAsymptotic = (N > 100))
}
\arguments{
\item{p}{the p-value for which the quantile is computed}

\item{N}{the length of residual vector considered in the normalized 3-depth.
This value is only relevant if \code{useAsymptotics=FALSE}}

\item{useAsymptotics}{a boolean variable. If \code{TRUE}, asymptotic
quantiles from the package \code{rexpar} are used (not published on CRAN,
only available via GitHub). Otherwise, simulated quantiles for the specified
N are used.}
}
\value{
the p-quantile of the normalized 3-depth
}
\description{
The (asymptotic) quantiles of the normalized 3-depth if the signs of the
residuals are i.i.d. unifomly in {-1,1} distributed random variables. By
default, asymptotic quantiles are used for residuals of length larger than 100.
Otherwise, simulated quantiles from an unpublished work by M. Horn
and C.H.Müller (2019) are used.
}
\seealso{
\code{\link{normed3depth}}
}
