% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoExp.R
\name{qhypoexp2}
\alias{qhypoexp2}
\title{Quantile function of the hypoexponential distribution}
\usage{
qhypoexp2(p, rates, searchInterval = c(0, 1e+10))
}
\arguments{
\item{p}{the p-value at which the function is evaluated}

\item{rates}{a vector of rates used in the hypoexponential distribution}

\item{searchInterval}{an initial search interval for \code{\link{uniroot}}}
}
\value{
the quantile at \code{p}
}
\description{
An alternative calculation for the quantile function of the hypoexponential 
distribution based on a matrix representation. A faster but numerically 
more instable implementation for the cdf is given in \code{sdprisk::qhypoexp}
}
\seealso{
\code{\link{phypoexp2}, \link{dhypoexp2}}
}
