% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{sample_w}
\alias{sample_w}
\title{Simulation of exponentially distributed random variable with outliers}
\usage{
sample_w(rates, distr = rexp, withOutliers = FALSE,
  outlierExceptions = NULL, LowerOutlierChance = 0.1,
  LowerOutlierFunc = function(x) {     return(1) },
  UpperOutlierChance = 0.1, UpperOutlierFunc = function(x) {    
  return(10 * x) })
}
\arguments{
\item{rates}{a vector of rates. The i-th coordinate of the vector determines
the rate used for the rate of the i-th entry of the resulting sample vector}

\item{distr}{a sample distribution with two arguments: the number N of samples
and the N-dimensional vector \code{rates} of rates (cf. exponential distribution)}

\item{withOutliers}{a boolean variable. If set \code{TRUE}, the data is
contaminated with outliers created using the functions \code{LowerOutlierFunc}
and \code{UpperOutlierFunc}.}

\item{outlierExceptions}{an optional vector of positions in the resulting
vector which should never be altered by the outlier functions.}

\item{LowerOutlierChance, UpperOutlierChance}{the probabilities to apply the
functions \code{LowerOutlierFunc} and \code{UpperOutlierFunc} to create
outliers.}

\item{LowerOutlierFunc, UpperOutlierFunc}{functions used to create outliers.}
}
\value{
a vector of exponentially distributed random variables. If
\code{withOutliers} is \code{TRUE}, each random variable has a chance to be
 altered using the functions \code{LowerOutlierFunc} and \code{UpperOutlierFunc}
}
\description{
Samples a vector of (by default) exponentially distributed random variables with rates
given in the vector \code{rates}. The sample can be contaminated with outliers
given as an alteration of single exponential samples by either the function
\code{LowerOutlierFunc} or by the function \code{UpperOutlierFunc}. A
contamination occurs with probability \code{LowerOutlierChance+UpperOutlierChance}
whereas the choice between \code{LowerOutlierFunc} and \code{UpperOutlierFunc}
is made with probabilities proportional to \code{LowerOutlierChance} and
\code{UpperOutlierChance}. The sample distribution can be altered by changing
\code{distr}. Note that the prediction methods in this paper are designed for
the exponential distribution only.
}
\seealso{
\code{\link{singleSimulation}}
}
