% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConBandsExp.R
\name{sampleStandQuants}
\alias{sampleStandQuants}
\title{Sample of standardized quantiles}
\usage{
sampleStandQuants(theta, s, Iv, I, linkfun = h_basq)
}
\arguments{
\item{theta}{the true model parameter in the underlying model. Note that
the distribution of the resulting quantiles does not depend on \code{theta}
if the default \code{linkfun} is used. Hence the input is irrelevant in
the default setting.}

\item{s, Iv, I}{parameter settings in the experiment.}

\item{linkfun}{link function used for the rate. Default choice is
\code{\link{h_basq}}.}
}
\value{
a vector of sample quantiles of the standardized exponential distribution.
}
\description{
Computes sample quantiles of standardized exponentially distributed random
variables. The standardization works as follows: First compute the MLE for
\code{theta} based on the samples; see \code{\link{estML}} for the computation
of the MLE. Then compute the estimated rates by using the MLE for \code{theta}
in the model; see \code{\link{getRates}}. The standardized exponentially distributed
random variables then are defined as the product of the samples with their
estimated rates. Note that this yields asymptotically standard exponentially distributed
random variables if the estimation for \code{theta} is consistent.
}
\seealso{
\code{\link{getRates}, \link{h_basq}, \link{estML}, \link{getConBands}}
}
