% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{simulateKRepetition}
\alias{simulateKRepetition}
\title{Repetition of \code{\link{singleSimulation}} with an increasing number of observations}
\usage{
simulateKRepetition(K_range, theta, block_s, block_Iv, block_I, pred_s,
  pred_start, pred_pos, pred_I, alpha, alpha1Func, distr, withOutliers,
  LowerOutlierChance, LowerOutlierFunc, UpperOutlierChance,
  UpperOutlierFunc, methods)
}
\arguments{
\item{K_range}{a vector of integers. Each element K of the vector
yields an application \code{\link{singleSimulation}} in which the length
of the observed data is proportional to K.}

\item{theta, pred_s, pred_start, pred_pos, pred_I, distr, withOutliers, LowerOutlierChance, LowerOutlierFunc, UpperOutlierChance, UpperOutlierFunc, methods}{parameter
choices in \code{\link{singleSimulation}}}

\item{block_s, block_Iv, block_I}{choices which the parameters of
the observed data are based on.}

\item{alpha}{the overall alpha-error in the prediction. Used in \code{\link{singleSimulation}}}

\item{alpha1Func}{a function on alpha and K. The return value is the
\code{alpha1}-value used in \code{\link{singleSimulation}}. The
\code{alpha2}-value in \code{\link{singleSimulation}} is set to be
\code{1-(1-alpha)/(1-alpha1)}}
}
\value{
a three-dimensional array \code{result} that contains the average performance of
each prediction method for each value of K in \code{K_range}.
The i-th row of the matrix contains the average coverage rate, length, and
interval score of the method \code{methods[i]}. The entry \code{result[ , ,K]}
is the return value of \code{\link{singleSimulation}} in the case
where the observed data is a \code{K_range[K]}-repetition of the parameters
\code{block_s, block_Iv, block_I}
}
\description{
A function to apply \code{\link{singleSimulation}} several times
to data sets where \code{obs_s, obs_Iv, obs_I} are K-times
repetitions of the choice for \code{block_s, block_Iv, block_I}
for every K in \code{K_range}.
}
\seealso{
\code{\link{singleSimulation}}
}
