% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{singleSimulation}
\alias{singleSimulation}
\title{Single simulation}
\usage{
singleSimulation(theta, obs_s, obs_Iv, obs_I, pred_s, pred_start, pred_pos,
  pred_I, alpha, alpha1, alpha2, distr, withOutliers, LowerOutlierChance,
  LowerOutlierFunc, UpperOutlierChance, UpperOutlierFunc, methods)
}
\arguments{
\item{theta}{a fixed model parameter used to determine the rates
of waiting times via \code{\link{getRates}} (in the Basquin Model)}

\item{obs_s, obs_Iv, obs_I}{parameters for the observed data. The observed
waiting times are exponentially distributed random variables with
rates obtained via \code{\link{getRates}}}

\item{pred_s, pred_start, pred_pos, pred_I}{parameter settings for the
experiment that is to be predicted. If \code{pred_start>1}, then
the first \code{pred_start-1} waiting times of the new experiment are
part of the observed data. The vector \code{pred_pos} contains the indices
of jump times that are to be predicted.}

\item{alpha, alpha1, alpha2}{error choices for the predictions; see
\code{\link{predfail}} for more details}

\item{distr, withOutliers, LowerOutlierChance, LowerOutlierFunc, UpperOutlierChance, UpperOutlierFunc}{options
to add outliers to the data; see \code{\link{sample_w}} for more details}

\item{methods}{a vector of methods used for predictions. These methods
need to be valid choices in \code{\link{predfail}}}
}
\value{
a matrix that contains the average performance of each prediction method.
The i-th row of the matrix contains the average coverage rate, length, and
interval score of the method \code{methods[i]}
}
\description{
Perform a single simulation with specified parameters. The simulations
samples observations according to the parameters \code{obs_s,obs_Iv,obs_I}.
Based on the observation, predictions are produced using the
specified \code{methods}. Observed data can be contaminated with outliers
\code{withOutliers=TRUE}.
}
\seealso{
\code{\link{sample_w}, \link{predfail}}
}
