% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConBandsExp.R
\name{testExpDist}
\alias{testExpDist}
\title{Test on exponential assumption}
\usage{
testExpDist(data, alpha, linkfun = h_basq, increaseZeros = TRUE,
  zeroReplacement = 1, n_sim = 1e+06, makePlot = TRUE,
  q_type = "l", col_band = "gray", fill_bands = TRUE,
  fill_col = "gray", xlab = "Quantiles of Exp(1)",
  ylab = "Sample Quantiles", main = paste("QQ-Plot with ", 100 * (1 -
  alpha), "\% confidence band", sep = ""))
}
\arguments{
\item{data}{a list storing four vectors: the waiting times \code{w},
the stresses \code{s} in each experiment, the number of failures \code{Iv}
in each experiment, the total number \code{I} of components in each experiment.}

\item{alpha}{the error chosen for the confidence bands computed via \code{\link{getConBands}}}

\item{linkfun}{the link function for the rates in the model. Default is the
Basquin link function \code{\link{h_basq}}.}

\item{increaseZeros}{a boolean variable. If \code{TRUE}, then all zeros among
the waiting times in \code{data} are replaced by the value \code{zeroReplacement}}

\item{zeroReplacement}{a positive number as a replacement for all zeros in
among the waiting times in \code{data}}

\item{n_sim}{the number of repetitions in the simulation of the confidence
band, see \code{\link{getConBands}}.}

\item{makePlot}{a boolean variable. If true, the function produces a QQ-plot
with a confidence band via \code{\link{plotConBands}}.
\code{col_band} and \code{fill_col} provide the colors for the border of the
confidence band and the fill color of the bands. The fill color is only used if
\code{fill_bands} is \code{TRUE}.
\@param q_type,cold_band,fill_bands,fill_col,xlab,ylab,main parameter choices
in \code{\link{plotConBands}}.}
}
\value{
a list with three entries: the first entry is a boolean variable
\code{rejectNull} which is \code{TRUE} if the test rejects the Null
Hypothesis, i.e. if the exponential assumption is rejected. The
second entry is a vector \code{pos_outside} which provides the positions
in the sample quantile vector which lie outside the confidence band. The
last entry \code{distance} yields the (directed) distances of the sample quantiles (with
index in \code{pos_outside}) to the confidence band. Negative distances
correspond to quantiles below the lower bound of the confidence band.
}
\description{
A \code{alpha}-level significance test to determine whether the waiting times
of the given \code{data} follow exponential distributions with rates
determined via the link function \code{linkfun}, see \code{\link{getRates}}.
The test also produces a Q-Q-plot with a simultaneous \code{(1-alpha)}-confidence
band if \code{makePlot} is \code{TRUE}. Since prediction intervals based on
\code{\link{predfail}} assume that the waiting times follow exponential distribution,
we recommend using this test before applying \code{\link{predfail}}.
Note that the confidence bands are computed via simulation hence causing
this function to have a slow runtime. Reducing \code{n_sim} speeds up the
function but increases the approximation error due to the simulation.
Note that due to discrete time measurements and the continuity of the
exponential distribution, the test would always reject if any of the waiting
times equal zero. We thus recommend to increase zero values slightly, e.g.
by setting them to one instead (this is only recommended if the typical
value of a waiting time is way above one). This will be done if \code{increaseZeros}
is \code{TRUE}. In that case, all zeros are replaced by \code{zeroReplacement}.
}
\seealso{
\code{\link{getConBands},\link{plotConBands},\link{predfail}}
}
