% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationStudy.R
\name{whiteNoiseSampler}
\alias{whiteNoiseSampler}
\title{Generate sample distribution with white noise}
\usage{
whiteNoiseSampler(sd = 10000, min.val = 1)
}
\arguments{
\item{sd}{the standard deviation used for the white noise}

\item{min.val}{the minimum value for the sampled random variabe, usually to
ensure nonnegativity. Default is set to 1.}
}
\value{
a function with arguments \code{N} and \code{rates} that generates
\code{N} random variables based on an exponential distribution with white noise
}
\description{
Generates a sample distribution with two arguments \code{N} and \code{rates}.
The distribution generates \code{N} random variables consisting of a
sum of exponentially distributed random variables with \code{rate=rates}
and normally distributed random variables with \code{mean} zero and a
standard deviation given by the parameter \code{sd}.
}
\seealso{
\code{\link{sample_w}}
}
